refactor(identity): application mock import 경로 정리 #48
